requirements.txt
albumentations==1.3.0
numpy==1.20.3
opencv_python_headless==4.5.4.60
pandas==1.3.4
Pillow==10.0.1
pytorch_metric_learning==1.0.0
scikit_learn==1.3.0
torch==1.10.0
tqdm==4.62.3
